\name{pronouncing-package}
\alias{pronouncing-package}
\alias{pronouncing}
\docType{package}
\title{
\packageTitle{pronouncing}
}
\description{
\packageDescription{pronouncing}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pronouncing}
\packageIndices{pronouncing}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{pronouncing}

Maintainer: \packageMaintainer{pronouncing}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
